IEEE-205 move Filter.module.scss under global_css
