+++ hackathon_site/dashboard/frontend/src/components/global_css/Filter.module.scss
@@ -1,5 +1,5 @@
-@import "./../assets/abstracts/variables";
-@import "./../assets/abstracts/mixins";
+@import "./../../assets/abstracts/variables";
+@import "./../../assets/abstracts/mixins";
 
 .filter {
     height: 100%;
