+++ hackathon_site/dashboard/frontend/src/components/orders/OrdersFilter/OrderFilter.tsx
@@ -11,7 +11,7 @@ import RadioGroup from "@material-ui/core/RadioGroup";
 import FormGroup from "@material-ui/core/FormGroup";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
 import { OrdersOrdering, OrderStatus } from "api/types";
-import styles from "../../Filter.module.scss";
+import styles from "../../global_css/Filter.module.scss";
 
 type OrderByOptions = {
     value: OrdersOrdering;
